make projects clean architecture
