+++ http/router/router.go
@@ -7,15 +7,10 @@ import (
 )
 
 func Init() *mux.Router {
-	// cfg := config.App
 	router := mux.NewRouter()
 
 	// Set up routes
-	router.HandleFunc("/convert", controllers.ConvertCurrencyHandler).Methods("POST")
-
-	// Start the server
-	// fmt.Println("Server listening on port 8080...")
-	// http.ListenAndServe(":" + cfg.Server.Port, router)
+	router.HandleFunc("/convert", controllers.ConvertCurrencyHandler).Methods("GET")
 
 	return router
 }
